Squashed commit of the following:
commit 862be958a9bca7165b4bfdb1731cece630ed9ce0
Merge: b062ef9 aacceeb
Author: Jon Shaw <jonshaw199@gmail.com>
Date:   Sat Nov 19 10:27:33 2022 -0800

    Merge branch 'main' into dam

commit b062ef9aafb275752933c9ba0ee551f9ecef8176
Merge: e0b12f8 2a24127
Author: Jon Shaw <jonshaw199@gmail.com>
Date:   Sat Nov 19 09:52:13 2022 -0800

    Merge branch 'main' into dam

commit e0b12f8461ff0369b4c58d340125cd37ca12bbb5
Author: Jon Shaw <jonshaw199@gmail.com>
Date:   Thu Nov 17 21:30:49 2022 -0800

    bugfixes

commit 134bd5cbd6fc3d5f1d4404bb5c60d0f1d1d4473d
Author: Jon Shaw <jonshaw199@gmail.com>
Date:   Thu Nov 17 21:10:09 2022 -0800

    String device IDs (unique; will come from DB); orgID can be determined by senderID

+++ firmware/src/stateEnt/pattern/noisePlusPalette/noisePlusPalette.cpp
@@ -10,8 +10,8 @@
 #define COLOR_ORDER GRB
 
 // Params for width and height
-const uint8_t kMatrixWidth = CNT;
-const uint8_t kMatrixHeight = 1;
+const int kMatrixWidth = CNT;
+const int kMatrixHeight = 1;
 
 // Param for different pixel layouts
 const bool kMatrixSerpentineLayout = true;
@@ -436,7 +436,7 @@ uint16_t NoisePlusPalette::XY(uint8_t x, uint8_t y)
     if (y & 0x01)
     {
       // Odd rows run backwards
-      uint8_t reverseX = (kMatrixWidth - 1) - x;
+      int reverseX = (kMatrixWidth - 1) - x;
       i = (y * kMatrixWidth) + reverseX;
     }
     else
